Cleaning up the ng-add command a bit
