+++ src/schematics/ng-add.ts
@@ -1,66 +1,11 @@
-import { SchematicContext, SchematicsException, Tree } from '@angular-devkit/schematics';
-import { NodePackageInstallTask, RunSchematicTask } from '@angular-devkit/schematics/tasks';
-import { listProjects, projectPrompt, projectTypePrompt } from './utils';
-import { Workspace } from './interfaces';
+import { chain, Tree } from '@angular-devkit/schematics';
+import { listProjects, projectPrompt, getWorkspace, getProject } from './utils';
 import { DeployOptions, NgAddNormalizedOptions } from './ng-add-common';
 import { addFirebaseFunctionsDependencies, setupUniversalDeployment } from './ng-add-ssr';
 import { addFirebaseHostingDependencies, setupStaticDeployment } from './ng-add-static';
 
-function getWorkspace(
-  host: Tree
-): { path: string; workspace: Workspace } {
-  const possibleFiles = ['/angular.json', '/.angular.json'];
-  const path = possibleFiles.filter(p => host.exists(p))[0];
-
-  const configBuffer = host.read(path);
-  if (configBuffer === null) {
-    throw new SchematicsException(`Could not find angular.json`);
-  }
-
-  // We can not depend on this library to have be included in older (or newer) Angular versions.
-  // Require here, since the schematic will add it to the package.json and install it before
-  // continuing.
-  const { parse }: typeof import('jsonc-parser') = require('jsonc-parser');
-
-  const workspace = parse(configBuffer.toString()) as Workspace|undefined;
-  if (!workspace) {
-    throw new SchematicsException('Could not parse angular.json');
-  }
-
-  return {
-    path,
-    workspace
-  };
-}
-
-const getProject = (options: DeployOptions, host: Tree) => {
-  const { workspace } = getWorkspace(host);
-  const projectName = options.project || workspace.defaultProject;
-
-  if (!projectName) {
-    throw new SchematicsException(
-      'No Angular project selected and no default project in the workspace'
-    );
-  }
-
-  const project = workspace.projects[projectName];
-  if (!project) {
-    throw new SchematicsException(
-      'The specified Angular project is not defined in this workspace'
-    );
-  }
-
-  if (project.projectType !== 'application') {
-    throw new SchematicsException(
-      `Deploy requires an Angular project type of "application" in angular.json`
-    );
-  }
-
-  return {project, projectName};
-};
-
 export const setupProject =
-  (host: Tree, options: DeployOptions & { isUniversalProject: boolean, firebaseProject: string }) => {
+  (host: Tree, options: DeployOptions & { firebaseProject: string, isUniversalProject: boolean }) => {
     const { path: workspacePath, workspace } = getWorkspace(host);
 
     const {project, projectName} = getProject(options, host);
@@ -78,50 +23,28 @@ export const setupProject =
         tree: host,
         project
       });
+    } else {
+      return setupStaticDeployment({
+        workspace,
+        workspacePath,
+        options: config,
+        tree: host,
+        project
+      });
     }
-    return setupStaticDeployment({
-      workspace,
-      workspacePath,
-      options: config,
-      tree: host,
-      project
-    });
-  };
+};
 
 export const ngAddSetupProject = (
-  options: DeployOptions & { isUniversalProject: boolean }
+  options: DeployOptions
 ) => async (host: Tree) => {
   const projects = await listProjects();
   const { firebaseProject } = await projectPrompt(projects);
-  return setupProject(host, {...options, firebaseProject});
+  const isUniversalProject = (global as any).setupAsAngularUniversalApp;
+  return setupProject(host, {...options, firebaseProject, isUniversalProject });
 };
 
-export const ngAdd = (options: DeployOptions) => (
-  host: Tree,
-  context: SchematicContext
-) => {
-
-  addFirebaseHostingDependencies(host, context);
-
-  const {project} = getProject(options, host);
-
-  // In Angular 12 it appears I might need some sort of timeout to allow
-  // node_modules to resolve?
-  const timeout = new Promise(resolve => setTimeout(resolve, 1_000));
-
-  return timeout.
-    then(() => projectTypePrompt(project)).
-    then(({ universalProject }: { universalProject: boolean }) => {
-      if (universalProject) {
-        addFirebaseFunctionsDependencies(host, context);
-      }
-      const projectOptions: DeployOptions & { isUniversalProject: boolean } = {
-        ...options,
-        isUniversalProject: universalProject
-      };
-      context.addTask(new RunSchematicTask('ng-add-setup-project', projectOptions), [
-        context.addTask(new NodePackageInstallTask())
-      ]);
-    }
-  );
-};
+export const ngAdd = (options: DeployOptions) => chain([
+  addFirebaseHostingDependencies(),
+  addFirebaseFunctionsDependencies(options),
+  ngAddSetupProject(options),
+]);
